scripts/requirements.txt

numpy==1.24.*
pyserial==3.5
cryptography==39.0.*
